requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
cryptography==44.0.3
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
django-oauth-toolkit==2.3.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
msgpack==1.1.0
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
requests==2.32.3
sorl-thumbnail==12.11.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
